+++ StarkBank/StarkBank/MerchantSession/MerchantSession.cs
@@ -64,7 +64,7 @@ namespace StarkBank
         }
 
 		public static IEnumerable<MerchantSession> Query(int? limit = null, DateTime? after = null, DateTime? before = null,
-            string status = null, List<string> tags = null, List<string> ids = null, User user = null)
+            string status = null, List<string> tags = null, List<string> ids = null, string holderId = null, User user = null)
         {
             (string resourceName, StarkCore.Utils.Api.ResourceMaker resourceMaker) = Resource();
             return Rest.GetList(
@@ -76,14 +76,15 @@ namespace StarkBank
                     { "before", new StarkCore.Utils.StarkDate(before) },
                     { "status", status },
                     { "tags", tags },
-                    { "ids", ids }
+                    { "ids", ids },
+                    { "holderId", holderId },
                 },
                 user: user
             ).Cast<MerchantSession>();
         }
 
         public static (List<MerchantSession> page, string pageCursor) Page(string cursor = null, int? limit = null, DateTime? after = null,
-            DateTime? before = null, string status = null, List<string> tags = null, List<string> ids = null, User user = null)
+            DateTime? before = null, string status = null, List<string> tags = null, List<string> ids = null, string holderId = null, User user = null)
         {
             (string resourceName, StarkCore.Utils.Api.ResourceMaker resourceMaker) = Resource();
             (List<StarkCore.Utils.SubResource> page, string pageCursor) = Rest.GetPage(
@@ -96,7 +97,8 @@ namespace StarkBank
                     { "before", new StarkCore.Utils.StarkDate(before) },
                     { "status", status },
                     { "tags", tags },
-                    { "ids", ids }
+                    { "ids", ids },
+                    { "holderId", holderId },
                 },
                 user: user
             );